add build step + ci
