+++ .gitignore
@@ -3,3 +3,4 @@
 /node_modules
 /npm-debug*
 /yarn-error*
+lib/svg-to-js*
