--- .travis.yml
@@ -1,3 +1,0 @@
-language: node_js
-node_js:
-  - "10"
