+++ package.json
@@ -2,31 +2,49 @@
   "name": "@nrk/svg-to-js",
   "version": "2.1.1",
   "description": "Module for concatenating SVG files into JavaScript",
-  "main": "index.js",
-  "scripts": {
-    "publish:patch": "npm version patch -m 'Release patch %s' && npm run push",
-    "publish:minor": "npm version minor -m 'Release minor %s' && npm run push",
-    "publish:major": "npm version major -m 'Release major %s' && npm run push",
-    "push": "git push && git push --tags && npm publish --access public",
-    "test": "jest && standard -v"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/nrkno/svg-to-js.git"
-  },
-  "keywords": [
-    "SVG",
-    "JavaScript",
-    "JS"
-  ],
   "author": "NRK <opensource@nrk.no> (https://www.nrk.no/)",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/nrkno/svg-to-js/issues"
   },
   "homepage": "https://github.com/nrkno/svg-to-js#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/nrkno/svg-to-js.git"
+  },
+  "keywords": [
+    "svg",
+    "javascript",
+    "js"
+  ],
+  "main": "lib/svg-to-js.cjs.js",
+  "module": "lib/svg-to-js.mjs",
+  "files": [
+    "lib/svg-to-js.cjs.js",
+    "lib/svg-to-js.mjs"
+  ],
+  "scripts": {
+    "publish:patch": "npm version patch -m 'Release patch %s' && npm run push",
+    "publish:minor": "npm version minor -m 'Release minor %s' && npm run push",
+    "publish:major": "npm version major -m 'Release major %s' && npm run push",
+    "postversion": "npm run build && git push && git push --tags && npm publish --access public",
+    "build": "rollup --config",
+    "test": "jest && standard -v"
+  },
   "devDependencies": {
     "jest": "24.9.0",
-    "standard": "14.1.0"
+    "rollup": "^1.27.8",
+    "rollup-plugin-buble": "^0.19.8",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-node-resolve": "^5.2.0",
+    "rollup-plugin-uglify": "^6.0.3",
+    "standard": "^14.1.0"
+  },
+  "standard": {
+    "globals": [
+      "describe",
+      "expect",
+      "it"
+    ]
   }
 }
